refactor(dashboard-api): one file per management handler
The seven were scaffolded into a single management.go, which is not how
the rest of this directory is organised -- admin_users_delete.go,
template_get.go, teams_list.go are all one operation each. Splitting
them means a PR implementing one handler touches one file, and the
surface is legible from a directory listing.

Named resource-first to match the existing files, so the project
operations sort together rather than by verb.

sendNotImplemented moves to utils_management.go, following the utils_
convention already here. It loses a caller with every handler that lands
and disappears with the last one.

management_test.go becomes management_contract_test.go: it never tested
management.go, it pins the wire shapes the caller depends on.
