+++ packages/dashboard-api/internal/handlers/management_project_member_delete.go
@@ -1,0 +1,12 @@
+package handlers
+
+import (
+	"github.com/gin-gonic/gin"
+
+	"github.com/e2b-dev/infra/packages/dashboard-api/internal/api"
+)
+
+// ManagementDeleteProjectMember removes a project member.
+func (s *APIStore) ManagementDeleteProjectMember(c *gin.Context, _ api.TeamID, _ api.UserId) {
+	sendNotImplemented(c)
+}
